// web/src/types/windi.ts
export type WindiUtilities = {
  font?: string
  text?: string
  underline?: string
  list?: string
  bg?: string
  gradient?: string
  border?: string
  divide?: string
  ring?: string
  icon?: string
  container?: string
  p?: string
  m?: string
  space?: string
  w?: string
  minW?: string
  maxW?: string
  h?: string
  minH?: string
  maxH?: string
  flex?: string
  grid?: string
  table?: string
  order?: string
  align?: string
  justify?: string
  place?: string
  display?: string
  pos?: string
  box?: string
  caret?: string
  isolation?: string
  object?: string
  overflow?: string
  overscroll?: string
  z?: string
  shadow?: string
  opactiy?: string
  blend?: string
  filter?: string
  backdrop?: string
  transition?: string
  animate?: string
  transform?: string
  appearance?: string
  cursor?: string
  outline?: string
  pointer?: string
  resize?: string
  select?: string
  sr?: string
}